+++ docs/source/retry-policy/default.md
@@ -12,7 +12,7 @@ To use in `Session`:
 # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 use scylla::{Session, SessionBuilder};
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retries::DefaultRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Box::new(DefaultRetryPolicy::new()))
@@ -37,7 +37,7 @@ To use in a [simple query](../queries/simple.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::query::Query;
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retries::DefaultRetryPolicy;
 
 // Create a Query manually and set the retry policy
 let mut my_query: Query = Query::new("INSERT INTO ks.tab (a) VALUES(?)");
@@ -66,7 +66,7 @@ To use in a [prepared query](../queries/prepared.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::prepared_statement::PreparedStatement;
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retries::DefaultRetryPolicy;
 
 // Create PreparedStatement manually and set the retry policy
 let mut prepared: PreparedStatement = session
