+++ docs/source/retry-policy/fallthrough.md
@@ -11,7 +11,7 @@ To use in `Session`:
 # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 use scylla::{Session, SessionBuilder};
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::FallthroughRetryPolicy;
+use scylla::execution::retries::FallthroughRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Box::new(FallthroughRetryPolicy::new()))
@@ -35,7 +35,7 @@ To use in a [simple query](../queries/simple.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::query::Query;
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::FallthroughRetryPolicy;
+use scylla::execution::retries::FallthroughRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Box::new(FallthroughRetryPolicy::new()))
@@ -61,7 +61,7 @@ To use in a [prepared query](../queries/prepared.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::prepared_statement::PreparedStatement;
 use scylla::transport::ExecutionProfile;
-use scylla::transport::retry_policy::FallthroughRetryPolicy;
+use scylla::execution::retries::FallthroughRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Box::new(FallthroughRetryPolicy::new()))
