+++ scylla/src/execution/mod.rs
@@ -1,5 +1,6 @@
 mod execution_profile;
 pub mod load_balancing;
+pub mod retries;
 pub mod speculative_execution;
 
 pub(crate) use execution_profile::ExecutionProfileInner;
