+++ scylla/src/execution/retries/downgrading_consistency_rp.rs
@@ -4,10 +4,10 @@ use scylla_cql::{
 };
 use tracing::debug;
 
-use crate::retry_policy::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
+use super::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
 
 /// Downgrading consistency retry policy - retries with lower consistency level if it knows\
-/// that the initial CL is unreachable. Also, it behaves as [DefaultRetryPolicy](crate::retry_policy::DefaultRetryPolicy)
+/// that the initial CL is unreachable. Also, it behaves as [DefaultRetryPolicy](crate::execution::retries::DefaultRetryPolicy)
 /// when it believes that the initial CL is reachable.
 /// Behaviour based on [DataStax Java Driver]\
 ///(<https://docs.datastax.com/en/drivers/java/3.11/com/datastax/driver/core/policies/DowngradingConsistencyRetryPolicy.html>)
