+++ scylla/src/execution/retries/mod.rs
@@ -1,0 +1,11 @@
+mod default_rp;
+mod downgrading_consistency_rp;
+mod fallthrough_rp;
+mod retry_policy;
+
+pub use default_rp::{DefaultRetryPolicy, DefaultRetrySession};
+pub use downgrading_consistency_rp::{
+    DowngradingConsistencyRetryPolicy, DowngradingConsistencyRetrySession,
+};
+pub use fallthrough_rp::{FallthroughRetryPolicy, FallthroughRetrySession};
+pub use retry_policy::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
