+++ scylla/src/session/session_test.rs
@@ -3,11 +3,11 @@ use crate::batch::{Batch, BatchStatement};
 use crate::cluster::metadata::Strategy::NetworkTopologyStrategy;
 use crate::cluster::metadata::{CollectionType, ColumnKind, CqlType, NativeType, UserDefinedType};
 use crate::cluster::Datacenter;
+use crate::execution::retries::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
 use crate::frame::response::result::Row;
 use crate::frame::value::ValueList;
 use crate::prepared_statement::PreparedStatement;
 use crate::query::Query;
-use crate::retry_policy::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
 use crate::sharding::Token;
 use crate::statement::Consistency;
 use crate::tracing::TracingInfo;
