+++ scylla/tests/integration/execution_profiles.rs
@@ -10,8 +10,8 @@ use scylla::query::Query;
 use scylla::statement::SerialConsistency;
 use scylla::{
     cluster::ClusterData,
+    execution::retries::{RetryPolicy, RetrySession},
     load_balancing::{LoadBalancingPolicy, RoutingInfo},
-    retry_policy::{RetryPolicy, RetrySession},
     speculative_execution::SpeculativeExecutionPolicy,
     test_utils::unique_keyspace_name,
     ExecutionProfile, SessionBuilder,
@@ -82,7 +82,7 @@ impl<const NODE: u8> LoadBalancingPolicy for BoundToPredefinedNodePolicy<NODE> {
 }
 
 impl<const NODE: u8> RetryPolicy for BoundToPredefinedNodePolicy<NODE> {
-    fn new_session(&self) -> Box<dyn scylla::retry_policy::RetrySession> {
+    fn new_session(&self) -> Box<dyn scylla::execution::retries::RetrySession> {
         self.report_node(Report::RetryPolicy);
         Box::new(self.clone())
     }
@@ -95,10 +95,10 @@ impl<const NODE: u8> RetryPolicy for BoundToPredefinedNodePolicy<NODE> {
 impl<const NODE: u8> RetrySession for BoundToPredefinedNodePolicy<NODE> {
     fn decide_should_retry(
         &mut self,
-        query_info: scylla::retry_policy::QueryInfo,
-    ) -> scylla::retry_policy::RetryDecision {
+        query_info: scylla::execution::retries::QueryInfo,
+    ) -> scylla::execution::retries::RetryDecision {
         self.report_consistency(query_info.consistency);
-        scylla::retry_policy::RetryDecision::DontRetry
+        scylla::execution::retries::RetryDecision::DontRetry
     }
 
     fn reset(&mut self) {}
