+++ scylla/tests/integration/lwt_optimisation.rs
@@ -1,6 +1,6 @@
 use crate::utils::test_with_3_node_cluster;
+use scylla::execution::retries::FallthroughRetryPolicy;
 use scylla::frame::types;
-use scylla::retry_policy::FallthroughRetryPolicy;
 use scylla::session::Session;
 use scylla::{frame::protocol_features::ProtocolFeatures, test_utils::unique_keyspace_name};
 use scylla::{ExecutionProfile, SessionBuilder};
